+++ CHANGES.rst
@@ -96,6 +96,7 @@ Unreleased
 -   Add new "edg" identifier for Edge in UserAgentPreparser.
     :issue:`1797`
 -   Upgrade the debugger to jQuery 3.5.1. :issue:`1802`
+-   Make MultiPart Form parser Sans-IO. :issue:`1788`
 
 
 Version 1.0.1